requirements.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
kafka-python
pydantic